requirements.txt
aws-cdk.aws_dms
aws-cdk.aws_ec2
aws-cdk.aws_sns
aws-cdk.aws_glue
aws-cdk.aws_lambda
aws-cdk.aws_sns_subscriptions
aws-cdk.aws_events_targets